requirements.txt
streamlit
pandas
numpy
Pillow
requests
openpyxl
streamlit-drawable-canvas
